ParallelUtilities.jl

ParallelUtilities.AbstractConstrainedProduct — Type

AbstractConstrainedProduct{T,N}

Supertype of ProductSplit and ProductSection.

ParallelUtilities.ProductSection — Type

ProductSection{T,N,Q}

Iterator that loops over a specified section of the outer product of the ranges provided in reverse-lexicographic order. The ranges need to be strictly increasing. Given N ranges, each element returned by the iterator will be a tuple of length N with one element from each range.

See also: ProductSplit

ParallelUtilities.ProductSection — Method

ProductSection(iterators::Tuple{Vararg{AbstractRange}}, inds::AbstractUnitRange)

Construct a ProductSection iterator that represents a 1D view of the outer product of the ranges provided in iterators, with the range of indices in the view being specified by inds.

Examples

julia> p = ParallelUtilities.ProductSection((1:3,4:6), 5:8);

julia> collect(p)
4-element Array{Tuple{Int64,Int64},1}:
 (2, 5)
 (3, 5)
 (1, 6)
 (2, 6)

julia> collect(p) == collect(Iterators.product(1:3, 4:6))[5:8]
true

ParallelUtilities.ProductSplit — Type

ProductSplit{T,N,Q}

Iterator that loops over the outer product of ranges in reverse-lexicographic order. The ranges need to be strictly increasing. Given N ranges, each element returned by the iterator will be a tuple of length N with one element from each range.

See also: ProductSection

ParallelUtilities.ProductSplit — Method

ProductSplit(iterators::Tuple{Vararg{AbstractRange}}, np::Integer, p::Integer)

Construct a ProductSplit iterator that represents the outer product of the iterators split over np workers, with this instance reprsenting the values on the p-th worker.

Examples

julia> ProductSplit((1:2,4:5), 2, 1) |> collect
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (2, 4)

julia> ProductSplit((1:2,4:5), 2, 2) |> collect
2-element Array{Tuple{Int64,Int64},1}:
 (1, 5)
 (2, 5)

Base.extrema — Method

extrema(ps::AbstractConstrainedProduct; dim::Integer)

Compute the extrema of the section of the range number dim contained in ps.

Examples

julia> ps = ProductSplit((1:2, 4:5), 2, 1);

julia> collect(ps)
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (2, 4)

julia> extrema(ps, dim = 1)
(1, 2)

julia> extrema(ps, dim = 2)
(4, 4)

Base.maximum — Method

maximum(ps::AbstractConstrainedProduct; dim::Integer)

Compute the maximum value of the section of the range number dim contained in ps.

Examples

julia> ps = ProductSplit((1:2,4:5),2,1);

julia> collect(ps)
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (2, 4)

julia> maximum(ps, dim = 1)
2

julia> maximum(ps, dim = 2)
4

Base.minimum — Method

minimum(ps::AbstractConstrainedProduct; dim::Integer)

Compute the minimum value of the section of the range number dim contained in ps.

Examples

julia> ps = ProductSplit((1:2, 4:5), 2, 1);

julia> collect(ps)
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (2, 4)

julia> minimum(ps, dim = 1)
1

julia> minimum(ps, dim = 2)
4

ParallelUtilities.childindex — Method

childindex(ps::AbstractConstrainedProduct, ind)

Return a tuple containing the indices of the individual AbstractRanges corresponding to the element that is present at index ind in the outer product of the ranges.

Note

The index ind corresponds to the outer product of the ranges, and not to ps.

Examples

julia> iters = (1:5, 2:4, 1:3);

julia> ps = ProductSplit(iters, 7, 1);

julia> ind = 6;

julia> cinds = ParallelUtilities.childindex(ps, ind)
(1, 2, 1)

julia> v = collect(Iterators.product(iters...));

julia> getindex.(iters, cinds) == v[ind]
true

See also: childindexshifted

ParallelUtilities.childindexshifted — Method

childindexshifted(ps::AbstractConstrainedProduct, ind)

Return a tuple containing the indices in the individual iterators given an index of a AbstractConstrainedProduct.

If the ranges (r1, r2, ...) are used to generate ps, then return (i1, i2, ...) such that ps[ind] == (r1[i1], r2[i2], ...).

Examples

julia> iters = (1:5, 2:4, 1:3);

julia> ps = ProductSplit(iters, 7, 3);

julia> psind = 4;

julia> cinds = ParallelUtilities.childindexshifted(ps, psind)
(3, 1, 2)

julia> getindex.(iters, cinds) == ps[psind]
true

See also: childindex

ParallelUtilities.dropleading — Method

dropleading(ps::AbstractConstrainedProduct)

Return a ProductSection leaving out the first iterator contained in ps. The range of values of the remaining iterators in the resulting ProductSection will be the same as in ps.

Examples

julia> ps = ProductSplit((1:5, 2:4, 1:3), 7, 3);

julia> collect(ps)
7-element Array{Tuple{Int64,Int64,Int64},1}:
 (5, 4, 1)
 (1, 2, 2)
 (2, 2, 2)
 (3, 2, 2)
 (4, 2, 2)
 (5, 2, 2)
 (1, 3, 2)

julia> ParallelUtilities.dropleading(ps) |> collect
3-element Array{Tuple{Int64,Int64},1}:
 (4, 1)
 (2, 2)
 (3, 2)

ParallelUtilities.extrema_commonlastdim — Method

extrema_commonlastdim(ps::AbstractConstrainedProduct)

Return the reverse-lexicographic extrema of values taken from ranges contained in ps, where the pairs of ranges are constructed by concatenating the ranges along each dimension with the last one.

For two ranges this simply returns ([first(ps)], [last(ps)]).

Examples

julia> ps = ProductSplit((1:3,4:7,2:7), 10, 2);

julia> collect(ps)
8-element Array{Tuple{Int64,Int64,Int64},1}:
 (3, 6, 2)
 (1, 7, 2)
 (2, 7, 2)
 (3, 7, 2)
 (1, 4, 3)
 (2, 4, 3)
 (3, 4, 3)
 (1, 5, 3)

julia> extrema_commonlastdim(ps)
([(1, 2), (6, 2)], [(3, 3), (5, 3)])

ParallelUtilities.extremadims — Method

extremadims(ps::AbstractConstrainedProduct)

Compute the extrema of the sections of all the ranges contained in ps. Functionally this is equivalent to

map(i -> extrema(ps, dim = i), 1:ndims(ps))

but it is implemented more efficiently.

Returns a Tuple containing the (min,max) pairs along each dimension, such that the i-th index of the result contains the extrema along the section of the i-th range contained locally.

Examples

julia> ps = ProductSplit((1:2, 4:5), 2, 1);

julia> collect(ps)
2-element Array{Tuple{Int64,Int64},1}:
 (1, 4)
 (2, 4)

julia> extremadims(ps)
((1, 2), (4, 4))

ParallelUtilities.indexinproduct — Method

indexinproduct(iterators::NTuple{N,AbstractRange}, val::NTuple{N,Any}) where {N}

Return the index of val in the outer product of iterators, where iterators is a Tuple of increasing AbstractRanges. Return nothing if val is not present.

Examples

julia> iterators = (1:4, 1:3, 3:5);

julia> val = (2, 2, 4);

julia> ind = ParallelUtilities.indexinproduct(iterators,val)
18

julia> collect(Iterators.product(iterators...))[ind] == val
true

ParallelUtilities.localindex — Method

localindex(ps::AbstractConstrainedProduct{T}, val::T) where {T}

Return the index of val in ps. Return nothing if the value is not found.

Examples

julia> ps = ProductSplit((1:3, 4:5:20), 3, 2);

julia> collect(ps)
4-element Array{Tuple{Int64,Int64},1}:
 (2, 9)
 (3, 9)
 (1, 14)
 (2, 14)

julia> localindex(ps, (3,9))
2

ParallelUtilities.nelements — Method

nelements(ps::AbstractConstrainedProduct; dim::Integer)

Compute the number of unique values in the section of the dim-th range contained in ps.

Examples

julia> ps = ProductSplit((1:5, 2:4, 1:3), 7, 3);

julia> collect(ps)
7-element Array{Tuple{Int64,Int64,Int64},1}:
 (5, 4, 1)
 (1, 2, 2)
 (2, 2, 2)
 (3, 2, 2)
 (4, 2, 2)
 (5, 2, 2)
 (1, 3, 2)

julia> ParallelUtilities.nelements(ps, dim = 1)
5

julia> ParallelUtilities.nelements(ps, dim = 2)
3

julia> ParallelUtilities.nelements(ps, dim = 3)
2

ParallelUtilities.ntasks — Function

ntasks(iterators::Tuple)

The total number of elements in the outer product of the ranges contained in iterators, equal to prod(length, iterators)

ParallelUtilities.nworkersactive — Method

nworkersactive(iterators::Tuple)

Number of workers required to contain the outer product of the iterators.

ParallelUtilities.pmapbatch — Method

pmapbatch(f, iterators, [mapargs...]; 
    [num_workers::Int = nworkersactive(iterators)], [mapkwargs...])

Evaluate the function f in parallel, where each worker gets a part of the entire parameter space sequentially. The argument iterators needs to be a strictly-increasing range, or a tuple of such ranges. The outer product of these ranges forms the entire range of parameters that is processed in batches on the workers. Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the function f.

Additionally, the number of workers to be used may be specified using the keyword argument num_workers. In this case the first num_workers available workers are used in the evaluation.

pmapbatch(f, T::Type, iterators, [mapargs...];
    [num_workers::Int = nworkersactive(iterators)], [mapkwargs...])

Evaluate f in parallel, and convert the returned value to type T. The method is type stable if T is concrete. Values returned by f will be type-coerced if possible, and an error will be raised otherwise.

See also: pmapreduce, pmapsum

ParallelUtilities.pmapbatch_elementwise — Method

pmapbatch_elementwise(f, iterators, [mapargs...]; 
    [num_workers::Int = nworkersactive(iterators)], [mapkwargs...])

Evaluate the function f in parallel, where each worker gets a part of the entire parameter space sequentially. The argument iterators needs to be a strictly-increasing range of intergers, or a tuple of such ranges. The outer product of these ranges forms the entire range of parameters that is processed elementwise by the function f. The individual tuples are splatted and passed as arguments to f. Given n ranges in iterators, the function f will receive n values at a time.

Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the function f. Additionally, the number of workers to be used may be specified using the keyword argument num_workers. In this case the first num_workers available workers are used in the evaluation.

See also: pmapbatch

ParallelUtilities.pmapreduce — Method

pmapreduce(fmap, freduce, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

Evaluate a parallel mapreduce over the range spanned by the outer product of the iterators.

iterators must be a strictly-increasing range of integers, or a tuple of such ranges. The outer product of the ranges is split evenly across the workers. The function fmap receives a ProductSplit iterator as its first argument that acts as a collection of tuples. One may index into a ProductSplit or iterate over one to access individual tuples of integers.

The reduction function freduce is expected to accept a collection of mapped values. Note that this is different from the standard mapreduce operation in julia that expects a binary reduction operator. For example, fmap should be sum and not +. In case a binary operator op is to be used in the reduction, one may pass it as Base.splat(op) or wrap it in an anonymous function as x -> op(x...).

Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the mapping function fmap.

pmapreduce(fmap, Tmap, freduce, Treduce, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

The types Tmap and Treduce are the return types of the map and reduce operations respectively. The returned values will be coerced to the specified types if possible, throwing an error otherwise.

Keyword Arguments

showprogress::Bool = false : Displays a progress-bar if set to true

progressdesc = "Progress in pmapreduce : " : Leading text in the progress-bar

See also: pmapreduce_commutative, pmapsum

ParallelUtilities.pmapreduce_commutative — Method

pmapreduce_commutative(fmap, freduce, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

Evaluate a parallel mapreduce over the range spanned by the outer product of the iterators. The operation is assumed to be commutative, results obtained may be incorrect otherwise.

The argument iterators must be a strictly-increasing range of integers, or a tuple of such ranges. The outer product of the ranges is split evenly across the workers. The function fmap receives a ProductSplit iterator as its first argument that acts as a collection of tuples. One may index into a ProductSplit or iterate over one to access individual tuples of integers.

The reduction function freduce is expected to accept a collection of mapped values. Note that this is different from the standard mapreduce operation in julia that expects a binary reduction operator. For example, freduce should be sum and not +. In case a binary operator op is to be used in the reduction, one may pass it as Base.splat(op) or wrap it in an anonymous function as x -> op(x...).

Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the mapping function fmap.

pmapreduce_commutative(fmap, Tmap, freduce, Treduce, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

The types Tmap and Treduce are the return types of the map and reduce operations respectively. The returned values will be coerced to the specified types if possible, throwing an error otherwise.

Keyword Arguments

showprogress::Bool = false : Displays a progress-bar if set to true

progressdesc = "Progress in pmapreduce : " : Leading text in the progress-bar

See also: pmapreduce_commutative_elementwise, pmapreduce, pmapsum

ParallelUtilities.pmapreduce_commutative_elementwise — Method

pmapreduce_commutative_elementwise(fmap, freduce, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

Evaluate a parallel mapreduce over the range spanned by the outer product of the iterators. The argument iterators must be a strictly-increasing range of integers, or a tuple of such ranges. The map is evaluated elementwise over the entire range of parameters. The reduction is assumed to be commutative, results obtained may be incorrect otherwise.

The reduction function freduce is expected to accept a collection of mapped values. Note that this is different from the standard mapreduce operation in julia that expects a binary reduction operator. For example, freduce should be sum and not +. In case a binary operator op is to be used in the reduction, one may pass it as Base.splat(op) or wrap it in an anonymous function as x -> op(x...).

Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the mapping function fmap.

pmapreduce_commutative_elementwise(fmap, Tmap, freduce, Treduce, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

The types Tmap and Treduce are the return types of the map and reduce operations respectively. The returned values will be coerced to the specified types if possible, throwing an error otherwise.

Keyword Arguments

showprogress::Bool = false : Displays a progress-bar if set to true

progressdesc = "Progress in pmapreduce : " : Leading text in the progress-bar

See also: pmapreduce_commutative

ParallelUtilities.pmapsum — Method

pmapsum(fmap, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

Evaluate a parallel mapreduce over the range spanned by the outer product of the iterators, where the reduction operation is a sum.

The argument iterators must be a strictly-increasing range of integers, or a tuple of such ranges. The outer product of the ranges is split evenly across the workers. The function fmap receives a ProductSplit iterator as its first argument that acts as a collection of tuples. One may index into a ProductSplit or iterate over one to access individual tuples of integers.

Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the mapping function fmap.

pmapsum(fmap, Tmap, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

The types Tmap is the return types of the map. The returned values will be coerced to the specified type if possible, throwing an error otherwise.

Keyword Arguments

showprogress::Bool = false : Displays a progress-bar if set to true

progressdesc = "Progress in pmapsum : " : Leading text in the progress-bar

See also: pmapreduce, pmapreduce_commutative

ParallelUtilities.pmapsum_elementwise — Method

pmapsum_elementwise(fmap, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

Evaluate a parallel mapreduce over the range spanned by the outer product of the iterators, where the reduction operation is a sum. The argument iterators must be a strictly-increasing range of integers, or a tuple of such ranges. The map is evaluated elementwise over the entire range of parameters.

Arguments mapargs and keyword arguments mapkwargs — if provided — are passed on to the mapping function fmap.

pmapsum_elementwise(fmap, Tmap, iterators, [mapargs...]; 
    <keyword arguments>, [mapkwargs...])

The type Tmap is the return type of the map. The returned values will be coerced to the specified type if possible, throwing an error otherwise.

Keyword Arguments

showprogress::Bool = false : Displays a progress-bar if set to true

progressdesc = "Progress in pmapreduce : " : Leading text in the progress-bar

See also: pmapreduce_commutative_elementwise, pmapsum

ParallelUtilities.procrange_recast — Method

procrange_recast(ps::AbstractConstrainedProduct, np_new::Integer)

Return the range of processor ranks that would contain the values in ps if the iterators used to construct ps were split across np_new processes.

Examples

julia> iters = (1:10, 4:6, 1:4);

julia> ps = ProductSplit(iters, 5, 2); # split across 5 processes initially

julia> procrange_recast(ps, 10) # If `iters` were spread across 10 processes
3:4

ParallelUtilities.procrange_recast — Method

procrange_recast(iterators::Tuple, ps::ProductSplit, np_new::Integer)

Return the range of processor ranks that would contain the values in ps if the outer produce of the ranges in iterators is split across np_new workers.

The values contained in ps should be a subsection of the outer product of the ranges in iterators.

Examples

julia> iters = (1:10, 4:6, 1:4);

julia> ps = ProductSplit(iters, 5, 2);

julia> procrange_recast(iters, ps, 10)
3:4

ParallelUtilities.sumcat_aligned — Method

sumcat_aligned(A::AbstractArray{T,N}...; dims) where {T,N}

Concatenate the arrays along the dimensions dims according to their axes, with overlapping sections being summed over. Returns an OffsetArray with the minimal axis span encompassing all the arrays.

dims may be an Integer or a collection of Integers, but all elements of dims must be from the range 1:N.

Examples

julia> ParallelUtilities.sumcat_aligned(ones(1:2), ones(4:5), dims=1)
5-element OffsetArray(::Array{Float64,1}, 1:5) with eltype Float64 with indices 1:5:
 1.0
 1.0
 0.0
 1.0
 1.0

julia> ParallelUtilities.sumcat_aligned(ones(1:2, 1:2), ones(2:3, 2:3), dims=(1,2))
3×3 OffsetArray(::Array{Float64,2}, 1:3, 1:3) with eltype Float64 with indices 1:3×1:3:
 1.0  1.0  0.0
 1.0  2.0  1.0
 0.0  1.0  1.0

julia> ParallelUtilities.sumcat_aligned(ones(1:2, 1:2), ones(3:4, 3:4), dims=(1,2))
4×4 OffsetArray(::Array{Float64,2}, 1:4, 1:4) with eltype Float64 with indices 1:4×1:4:
 1.0  1.0  0.0  0.0
 1.0  1.0  0.0  0.0
 0.0  0.0  1.0  1.0
 0.0  0.0  1.0  1.0

See also: sumhcat_aligned, sumvcat_aligned

ParallelUtilities.sumhcat_aligned — Method

sumhcat_aligned(A::AbstractArray{T,N}...) where {T,N}

Concatenate the arrays along the second dimension according to their axes, with overlapping sections being summed over. Returns an OffsetArray with the minimal axis span encompassing all the arrays.

The input arrays must be at least two-dimensional.

Examples

julia> ParallelUtilities.sumhcat_aligned(ones(2, 1:2), ones(2, 4:5))
2×5 OffsetArray(::Array{Float64,2}, 1:2, 1:5) with eltype Float64 with indices 1:2×1:5:
 1.0  1.0  0.0  1.0  1.0
 1.0  1.0  0.0  1.0  1.0

julia> ParallelUtilities.sumhcat_aligned(ones(1:2, 1:2), ones(1:2, 2:3))
2×3 OffsetArray(::Array{Float64,2}, 1:2, 1:3) with eltype Float64 with indices 1:2×1:3:
 1.0  2.0  1.0
 1.0  2.0  1.0

See also: sumcat_aligned, sumvcat_aligned

ParallelUtilities.sumvcat_aligned — Method

sumvcat_aligned(A::AbstractArray{T,N}...) where {T,N}

Concatenate the arrays along the first dimension according to their axes, with overlapping sections being summed over. Returns an OffsetArray with the minimal axis span encompassing all the arrays.

The input arrays must be at least one-dimensional.

Examples

julia> ParallelUtilities.sumvcat_aligned(ones(1:2), ones(4:5))
5-element OffsetArray(::Array{Float64,1}, 1:5) with eltype Float64 with indices 1:5:
 1.0
 1.0
 0.0
 1.0
 1.0

julia> ParallelUtilities.sumvcat_aligned(ones(1:2, 1:2), ones(2:3, 1:2))
3×2 OffsetArray(::Array{Float64,2}, 1:3, 1:2) with eltype Float64 with indices 1:3×1:2:
 1.0  1.0
 2.0  2.0
 1.0  1.0

See also: sumcat_aligned, sumhcat_aligned

ParallelUtilities.whichproc — Method

whichproc(iterators::Tuple, val::Tuple, np::Integer)

Return the processor rank that will contain val if the outer product of the ranges contained in iterators is split evenly across np processors.

Examples

julia> iters = (1:4, 2:3);

julia> np = 2;

julia> ProductSplit(iters, np, 2) |> collect
4-element Array{Tuple{Int64,Int64},1}:
 (1, 3)
 (2, 3)
 (3, 3)
 (4, 3)

julia> whichproc(iters, (2,3), np)
2

ParallelUtilities.whichproc_localindex — Method

whichproc_localindex(iterators::Tuple, val::Tuple, np::Integer)

Return (rank,ind), where rank is the rank of the worker that val will reside on if the outer product of the ranges in iterators is spread over np workers, and ind is the index of val in the local section on that worker.

Examples

julia> iters = (1:4, 2:8);

julia> np = 10;

julia> whichproc_localindex(iters, (2,4), np)
(4, 1)

julia> ProductSplit(iters, np, 4) |> collect
3-element Array{Tuple{Int64,Int64},1}:
 (2, 4)
 (3, 4)
 (4, 4)

ParallelUtilities.workersactive — Method

workersactive(iterators::Tuple)

Workers required to split the outer product of the iterators. If prod(length, iterators) < nworkers() then the first prod(length, iterators) workers are chosen.

ParallelUtilities.ClusterQueryUtils.gethostnames — Function

gethostnames(procs = workers())

Return the hostname of each worker in procs. This is obtained by evaluating Libc.gethostname() on each worker asynchronously.

Warn

gethostnames is deprecated in favor of hostnames

ParallelUtilities.ClusterQueryUtils.hostnames — Function

hostnames(procs = workers())

Return the hostname of each worker in procs. This is obtained by evaluating Libc.gethostname() on each worker asynchronously.

ParallelUtilities.ClusterQueryUtils.nodenames — Function

nodenames(procs = workers())

Return the unique hostnames that the workers in procs lie on. On an HPC system these are usually the hostnames of the nodes involved.

ParallelUtilities.ClusterQueryUtils.nprocs_node — Function

nprocs_node(procs = workers())

Return the number of workers on each host. On an HPC system this would return the number of workers on each node.

ParallelUtilities.ClusterQueryUtils.procs_node — Function

procs_node(procs = workers())

Return the worker ids on each host of the cluster. On an HPC system this would return the workers on each node.